Modifications to optimized DrawPolygonMesh. To use DrawPolygonMesh just type FASTMESH() inside PARAMS section. This optimizations are good for High polygon models (a few thousand faces) and work even better if the number of materials is keeped low (1 is perfect, after that performance goes down)
Issues:
1) Keep in mind this optimizations have high memory requirements.
There are a lot of buffers, so use it carefully.
2) First time the mesh is viewed it has to calculate the poly buffer,
so it takes 1-2 secs to draw. I think polybuffers can be precalc
before the mesh has to be drawn for 1st time, so perhaps that could
be a good way to avoid the "frozen seconds"

Known Bugs:
- The mesh have to be lighted, otherwise it crashes
- Dunno exactly why, but for high poly models fog doesn't
draws correctly, but it works perfect for my 1000 faces sphere :)


Known Bugs 2: The one that will make Jorrit kills me :)
I haven't updated the flush transforms, so it still
flushes the lightmap queues for every draw :|, but well
i had to meet the 0.94 deadline, so...

Unknown Bugs:
- As the title says, they're unknown... so i don't have
any idea, but i'm pretty sure they're there, waiting...