Changed more std::string instances to scfString.  Fixed a bug in the xml_defs that only existed b/c of a lame STL issue where std::string would fault it given a null pointer as an initializer.  Highly lame.
Also changed it so that all values in a definition file are now automation strings.  This means that they can be simple constant values like float, string, and integer -- or they can be dynamic function calls.  This is interesting and useful for several things, mostly because it allows definition files to get values from the running program.