Oktal reorganized the split swig interface files.
This should make it slightly easier to continue splitting the rest of
the modules out of cspace.i, now that cspace.i contains no code that is
common to all the modules. That common code is now in base*.i

It's also nice that there is no longer any common code replicated in all
the split module interface files.