First draft of the cseditor core framework in its (hopefully) final development iteration. This work is heavily based on Sueastside's code from the 'loading' SVN branch at r37819.
The main changes from the 'loading' branch are:
- More explicit access to the cseditor core component and to the creation of iEditor instances (see interface iEditorManager).
- The iEditor and its managers are no more accessed from the object registry.
- Reworked the initialization and registration of most editor components.
- Addition of the new iEditorComponent for components hidden from the GUI.
- Switch from listeners to the iEventQueue.
- The context is now split in several different interfaces.
- The loading feature is application dependent and has therefore been moved to a 'maploader' component instead.
- General cleaning of the API and of its method names.

The main missing features from the 'loading' branch are:
- the DAMN browser
- the settings manager
- the implementation of actions and operators is not yet complete
- some of the components are not yet integrated

Note also that there is clearly an issue with threading concurrency, with eg the objects not already available in the engine while the loading is supposed to be finished. The headers are not working neither for the same reason.

The API of the core cseditor framework is now mostly fixed, consider it probably as in the 'beta' status. Most of the remaining work has now to be made through additional editor components.