csdef: Retire CS-specific PRIx99 format macros for 64-bit datatypes.
Now that 64-bit machines are common, PRIx99 format macros for 64-bit
datatypes typically are platform-supplied. Consequently, drop the
CS-specific PRIx99 64-bit variations with the "CS_" prefix in favor of
the platform-supplied ones; and fake them up only on platforms which
lack them (but using the standardized names, not the "CS_"-prefixed
names).