openmp: Resolve "iteration variable is unsigned" warnings.
OpenMP 2.5 expects iteration variables to be signed and emits a diagnostic
if they are not. OpenMP 3.0 is more flexible and allows an unsigned, a
pointer type, or a constant-time random access iterator type.

Some platforms, such as OSX Lion, ship with OpenMP 2.5. Employ only
unsigned iteration variables in order to avoid annoying warnings on these
platforms.