vararg: iNativeWindow::Alert()
This vararg method triggers a compilation error[1] on OSX.  Work around
the problem by implementing the vararg method directly in the interface
as a thin inline non-virtual wrapper which delegates to the va_list
counterpart.

[1] Apple's llvm-g++ compiler has incomplete vararg support.  When it
encounters a virtual method using varargs in a virtual base class, it
aborts with error "generic thunk code fails for method foo::bar() which
uses '...'".